(1) Refactor, Change live display use heart sprite, enemy show sprite with remain count (2) New, Add new Font and Heart Sprite (3) Refactor, add map to store enemyType to sprite (4) Refactor, add current enemy list storage to metric and control (5) Refactor, each stage store enemy config to list and set Stage enemy size to metric and ctrl (6) Bug Fix, solve hero lives overflow problem

+++ app/src/main/java/com/example/overrun/enitities/character/CharacterBase.kt
@@ -23,7 +23,7 @@
         private var _speed : UInt = speed
         private var _eDir : eDirection = eDirection.eDOWN
         private var _etype : eCharacterType = etype
-        private var _lives : MutableState<UInt> = mutableStateOf(lives)
+        private var _lives : MutableState<Int> = mutableStateOf(lives.toInt())
         private var _finishedDie = false
 
         // GameObject Base owns collider
@@ -40,17 +40,17 @@
         public fun getSpeed() = _speed
 
         public fun setLives(lives: UInt) {
-            _lives.value = lives
+            _lives.value = lives.toInt()
         }
         public fun incrementLives(size: UInt){
-            _lives.value += size
+            _lives.value += size.toInt()
         }
         public fun decrementLives(size: UInt) {
-            _lives.value = (_lives.value - size).coerceAtLeast(0U)
+            _lives.value = (_lives.value - size.toInt()).coerceAtLeast(0)
         }
-        public fun getLives() = _lives.value.coerceAtLeast(0U)
+        public fun getLives() = _lives.value.coerceAtLeast(0)
 
-        public fun isDie() : Boolean = _lives.value <= 0U
+        public fun isDie() : Boolean = _lives.value <= 0
         public fun setDieFinished()
         {
             _finishedDie = true
